Few updates

+++ generate_changelog.sh
@@ -4,4 +4,6 @@ source config
 
 bash parse_tags.sh
 python3 generate_destination_changelog.py "$CHANGELOG_PATH" "$COMMIT_TAG" "$ISSUE_BASE_URL" "$DESTINATION_TEMPLATE" "$DESTINATION_FILE_NAME"
-rm "$CHANGELOG_PATH"
+if [ -d "$CHANGELOG_PATH" ]; then
+  rm "$CHANGELOG_PATH"
+fi
